Adding tick/attack cycle/attack rate to GameMonster
Also changing attack rate/cycle from a float to a long

+++ Assets/CrossMonster/Scripts/Monsters/IMonsterData.cs
@@ -8,6 +8,6 @@ namespace CrossMonsters {
         int GetDefenseType();
         int GetDamage();
         int GetDamageType();
-        float GetAttackRate();
+        long GetAttackRate();
     }
 }
